Extract RO-related calls to an RO service, allowing later swap to delius
Also simplify nested com structure - just com instead of com.com
