+++ server/index.js
@@ -28,13 +28,16 @@ const createNomisPushService = require('./services/nomisPushService')
 const createDeadlineService = require('./services/deadlineService')
 const createJobSchedulerService = require('./services/jobSchedulerService')
 const createNotificationJobs = require('./services/jobs/notificationJobs')
+const createNomisRoService = require('./services/nomisRoService')
 
 const signInService = createSignInService(audit)
 const licenceService = createLicenceService(licenceClient)
 const conditionsService = createConditionsService(config)
-const prisonerService = createPrisonerService(nomisClientBuilder)
+const nomisRoService = createNomisRoService(nomisClientBuilder)
+const roService = nomisRoService
+const prisonerService = createPrisonerService(nomisClientBuilder, roService)
 const caseListFormatter = createCaseListFormatter(logger, licenceClient)
-const caseListService = createCaseListService(nomisClientBuilder, licenceClient, caseListFormatter)
+const caseListService = createCaseListService(nomisClientBuilder, roService, licenceClient, caseListFormatter)
 const pdfService = createPdfService(logger, licenceService, conditionsService, prisonerService, pdfFormatter)
 const formService = createFormService(pdfService, pdfFormatter)
 const reportingService = createReportingService(audit)
