+++ server/services/prisonerService.js
@@ -1,12 +1,10 @@
 const logger = require('../../log.js')
 const { formatObjectForView } = require('./utils/formatForView')
-const { getIn, isEmpty } = require('../utils/functionalHelpers')
+const { getIn } = require('../utils/functionalHelpers')
 
 module.exports = { createPrisonerService }
 
-class RoRelationshipError extends Error {}
-
-function createPrisonerService(nomisClientBuilder) {
+function createPrisonerService(nomisClientBuilder, roService) {
   async function getPrisonerPersonalDetails(bookingId, token) {
     try {
       logger.info(`getPrisonerPersonalDetails: ${bookingId}`)
@@ -42,7 +40,7 @@ function createPrisonerService(nomisClientBuilder) {
       const [aliases, offences, com] = await Promise.all([
         nomisClient.getAliases(bookingId),
         nomisClient.getMainOffence(bookingId),
-        findResponsibleOfficer(bookingId, token),
+        roService.findResponsibleOfficer(bookingId, token),
       ])
 
       const { CRO, PNC } = selectEntriesWithTypes(await nomisClient.getIdentifiers(bookingId), ['PNC', 'CRO'])
@@ -55,7 +53,7 @@ function createPrisonerService(nomisClientBuilder) {
         PNC,
         offences,
         ...image,
-        com: [com],
+        com,
         aliases,
       })
     } catch (error) {
@@ -65,115 +63,7 @@ function createPrisonerService(nomisClientBuilder) {
   }
 
   async function getResponsibleOfficer(bookingId, token) {
-    logger.info(`getResponsibleOfficer: ${bookingId}`)
-    const com = await findResponsibleOfficer(bookingId, token)
-    return formatObjectForView({ com: [com] })
-  }
-
-  async function findResponsibleOfficer(bookingId, token) {
-    logger.info(`findResponsibleOfficer: ${bookingId}`)
-
-    try {
-      const nomisClient = nomisClientBuilder(token)
-      const roPersons = await getRoPersons(nomisClient, bookingId)
-
-      const allRoDetails = await Promise.all(
-        roPersons.map(roPerson => {
-          return getRoDetails(nomisClient, roPerson)
-        })
-      )
-
-      if (allRoDetails.length > 1) {
-        logger.warn(`Multiple RO relationships for bookingId: ${bookingId}`, allRoDetails)
-      }
-
-      return await selectRo(allRoDetails)
-    } catch (error) {
-      if (error instanceof RoRelationshipError) {
-        logger.error(`findResponsibleOfficer for: ${bookingId}`, error.message)
-        return { message: error.message }
-      }
-
-      logger.error(`findResponsibleOfficer for: ${bookingId}`, error.stack)
-      throw error
-    }
-  }
-
-  async function getRoPersons(nomisClient, bookingId) {
-    const relations = await getRoRelations(nomisClient, bookingId)
-    if (!relations[0]) {
-      throw new RoRelationshipError('No RO relationship')
-    }
-
-    const personIds = relations.filter(r => !isEmpty(r.personId))
-
-    if (isEmpty(personIds)) {
-      throw new RoRelationshipError('No RO person identifier')
-    }
-
-    return personIds
-  }
-
-  async function getRoRelations(nomisClient, bookingId) {
-    try {
-      return await nomisClient.getRoRelations(bookingId)
-    } catch (error) {
-      if (error.status === 404) {
-        logger.error(`RO relationship not found for booking id: ${bookingId}`)
-        throw new RoRelationshipError('No RO relationship')
-      }
-
-      logger.error('Error getting RO relationship', error.stack)
-      throw error
-    }
-  }
-
-  async function getRoDetails(nomisClient, roPerson) {
-    const { personId } = roPerson
-    const personIdentifiers = await getPersonIdentifiers(nomisClient, personId)
-
-    if (!personIdentifiers || personIdentifiers.length < 1) {
-      logger.warn(`No person identifiers for RO person: ${roPerson}`)
-      return roPerson
-    }
-
-    const id = personIdentifiers.find(record => record.identifierType === 'EXTERNAL_REL')
-
-    if (!id) {
-      logger.warn(`No EXTERNAL_REL person identifier for RO person: ${roPerson}`)
-      return roPerson
-    }
-
-    if (!id.identifierValue) {
-      logger.warn(`No EXTERNAL_REL person identifier value for RO person: ${roPerson}`)
-      return roPerson
-    }
-
-    return { ...roPerson, deliusId: id.identifierValue }
-  }
-
-  async function getPersonIdentifiers(nomisClient, personId) {
-    try {
-      return await nomisClient.getPersonIdentifiers(personId)
-    } catch (error) {
-      if (error.status === 404) {
-        logger.warn(`Person identifiers not found for person id: ${personId}`)
-        throw new RoRelationshipError('No RO external relationship')
-      }
-
-      logger.error('Error getting person identifiers', error.stack)
-      throw error
-    }
-  }
-
-  async function selectRo(allRoDetails) {
-    const roWithDeliusIds = allRoDetails.filter(r => !isEmpty(r.deliusId))
-
-    if (isEmpty(roWithDeliusIds)) {
-      throw new RoRelationshipError('No RO with a Delius staff code')
-    }
-
-    return roWithDeliusIds.sort((a, b) => b.personId - a.personId)[0]
+    return roService.findResponsibleOfficer(bookingId, token)
   }
 
   function getPrisonerImage(imageId, token) {
