+++ server/views/send/caToRo.pug
@@ -14,8 +14,8 @@ block content
 
       p The information will be submitted to:
 
-      if submissionTarget && submissionTarget.com
-        div#ro.bold.paddingBottom #{submissionTarget.com.name}
+      if submissionTarget
+        div#ro.bold.paddingBottom #{submissionTarget.name}
       else
         div.paddingBottom No responsible officer found
 
@@ -25,8 +25,8 @@ block content
         input(type="hidden" name="_csrf" value=csrfToken)
         input(type="hidden" name="bookingId" value=bookingId)
         input(type="hidden" name="transitionType" value='caToRo')
-        if submissionTarget && submissionTarget.com
-          input(type="hidden" name="submissionTarget" value=submissionTarget.com.name)
+        if submissionTarget
+          input(type="hidden" name="submissionTarget" value=submissionTarget.name)
 
         div.pure-u-1.inlineButtons.paddingBottom.smallPaddingTop
           input#continueBtn.requiredButton.button.button-start.midMarginBottom(type="submit" value="Submit address")
