+++ server/views/sent/caToRo.pug
@@ -4,8 +4,8 @@ block content
 
   include ./mixins/sentMixin
 
-  if submissionTarget && submissionTarget.com
-    +sentContent('Case sent', submissionTarget.com.name)
+  if submissionTarget
+    +sentContent('Case sent', submissionTarget.name)
   else
     +sentContent('Case sent', 'No responsible officer found')
