+++ test/healthcheck/healthcheckTest.js
@@ -161,22 +161,5 @@ describe('healthcheck', () => {
         expect(calledWith.checks.delius).to.eql(undefined)
       })
     })
-
-    it('should return unhealthy if delius rejects promise', () => {
-      const deliusCheckStubReject = sandbox.stub().rejects(404)
-
-      return healthcheckProxy(dbCheckStub, nomisApiCheckStub, deliusCheckStubReject, authCheckStub)(callback).then(
-        () => {
-          const calledWith = callback.getCalls()[0].args[1]
-
-          expect(callback).to.have.callCount(1)
-          expect(calledWith.healthy).to.eql(false)
-          expect(calledWith.checks.db).to.eql('OK')
-          expect(calledWith.checks.nomis).to.eql('OK')
-          expect(calledWith.checks.delius).to.eql(404)
-          expect(calledWith.checks.auth).to.eql('OK')
-        }
-      )
-    })
   })
 })
