+++ test/services/userAdminServiceTest.js
@@ -199,9 +199,9 @@ describe('userAdminService', () => {
     it('should lookup staff record for each unique assignedId', async () => {
       userClient.getCasesRequiringRo = sinon.stub().resolves([1, 2, 3])
       prisonerService.getResponsibleOfficer = sinon.stub()
-      prisonerService.getResponsibleOfficer.onCall(0).resolves({ com: { deliusId: 'delius0', name: 'deliusName0' } })
-      prisonerService.getResponsibleOfficer.onCall(1).resolves({ com: { deliusId: 'delius1', name: 'deliusName1' } })
-      prisonerService.getResponsibleOfficer.onCall(2).resolves({ com: { deliusId: 'delius1', name: 'deliusName2' } })
+      prisonerService.getResponsibleOfficer.onCall(0).resolves({ deliusId: 'delius0', name: 'deliusName0' })
+      prisonerService.getResponsibleOfficer.onCall(1).resolves({ deliusId: 'delius1', name: 'deliusName1' })
+      prisonerService.getResponsibleOfficer.onCall(2).resolves({ deliusId: 'delius1', name: 'deliusName2' })
 
       await service.getIncompleteRoUsers()
 
@@ -213,9 +213,9 @@ describe('userAdminService', () => {
     it('should add offender nomis for each incomplete', async () => {
       userClient.getCasesRequiringRo = sinon.stub().resolves([1, 2, 3])
       prisonerService.getResponsibleOfficer = sinon.stub()
-      prisonerService.getResponsibleOfficer.onCall(0).resolves({ com: { deliusId: 'delius0', name: 'deliusName0' } })
-      prisonerService.getResponsibleOfficer.onCall(1).resolves({ com: { deliusId: 'delius1', name: 'deliusName1' } })
-      prisonerService.getResponsibleOfficer.onCall(2).resolves({ com: { deliusId: 'delius2', name: 'deliusName2' } })
+      prisonerService.getResponsibleOfficer.onCall(0).resolves({ deliusId: 'delius0', name: 'deliusName0' })
+      prisonerService.getResponsibleOfficer.onCall(1).resolves({ deliusId: 'delius1', name: 'deliusName1' })
+      prisonerService.getResponsibleOfficer.onCall(2).resolves({ deliusId: 'delius2', name: 'deliusName2' })
       userClient.getRoUserByDeliusId = sinon.stub().resolves(null)
 
       await service.getIncompleteRoUsers()
@@ -228,9 +228,7 @@ describe('userAdminService', () => {
 
     it('should not return if present and onboarded', async () => {
       userClient.getCasesRequiringRo = sinon.stub().resolves([1])
-      prisonerService.getResponsibleOfficer = sinon
-        .stub()
-        .resolves({ com: { deliusId: 'delius1', name: 'deliusName1' } })
+      prisonerService.getResponsibleOfficer = sinon.stub().resolves({ deliusId: 'delius1', name: 'deliusName1' })
       userClient.getRoUserByDeliusId = sinon.stub().resolves({ onboarded: true })
 
       const result = await service.getIncompleteRoUsers()
@@ -240,9 +238,7 @@ describe('userAdminService', () => {
 
     it('should return mapped if present but not onboarded', async () => {
       userClient.getCasesRequiringRo = sinon.stub().resolves([1])
-      prisonerService.getResponsibleOfficer = sinon
-        .stub()
-        .resolves({ com: { deliusId: 'delius1', name: 'deliusName1' } })
+      prisonerService.getResponsibleOfficer = sinon.stub().resolves({ deliusId: 'delius1', name: 'deliusName1' })
       userClient.getRoUserByDeliusId = sinon.stub().resolves({ onboarded: false })
       nomisClient.getBooking = sinon.stub().resolves({ offenderNo: 'off1' })
 
@@ -262,9 +258,7 @@ describe('userAdminService', () => {
 
     it('should return unmapped if not present', async () => {
       userClient.getCasesRequiringRo = sinon.stub().resolves([1])
-      prisonerService.getResponsibleOfficer = sinon
-        .stub()
-        .resolves({ com: { deliusId: 'delius1', name: 'deliusName1' } })
+      prisonerService.getResponsibleOfficer = sinon.stub().resolves({ deliusId: 'delius1', name: 'deliusName1' })
       userClient.getRoUserByDeliusId = sinon.stub().resolves(null)
       nomisClient.getBooking = sinon.stub().resolves({ offenderNo: 'off1' })
